+++ src/components/index.js
@@ -1,7 +1,6 @@
-export { default as AutoLoginButton } from "./AutoLoginButton";
 export { default as Footer } from "./Footer";
 export { default as Header } from "./Header";
-export { default as JoinTypeButton } from "./JoinTypeButton";
+export { default as JoinTypeButton } from "./signUp/JoinTypeButton";
 export { default as ListThem } from "./ListThem";
 export { default as Pagination } from "./Pagination";
 export { default as SearchList } from "./SearchList";
