+++ src/constants/board.js
@@ -1,3 +1,5 @@
+import { faImage, faFile } from "@fortawesome/free-regular-svg-icons";
+
 const BOARD_NAMES = [
     {
         id: 1,
@@ -25,10 +27,24 @@ const TITLE_ITEMS = ["제목", "작성자", "작성일", "조회수", "좋아요
 
 const SEARCH_SELECT_ITEMS = ["전체기간", "게시물 + 작성자"];
 
+const FILE_IMAGE_BUTTON = [
+    {
+        id: 1,
+        faIcon: faImage,
+        text: "사진",
+    },
+    {
+        id: 2,
+        faIcon: faFile,
+        text: "파일",
+    },
+];
+
 const boardData = {
     BOARD_NAMES,
     TITLE_ITEMS,
     SEARCH_SELECT_ITEMS,
+    FILE_IMAGE_BUTTON,
 }
 
 export default boardData;
