--- src/constants/button.js
@@ -1,20 +1,0 @@
-import { faImage, faFile } from "@fortawesome/free-regular-svg-icons";
-
-const BUTTON_ITEMS = [
-    {
-        id: 1,
-        faIcon: faImage,
-        text: "사진",
-    },
-    {
-        id: 2,
-        faIcon: faFile,
-        text: "파일",
-    },
-];
-
-const buttonData = {
-    BUTTON_ITEMS,
-}
-
-export default buttonData;
