+++ src/constants/signUp.js
@@ -1,0 +1,14 @@
+export const CHECK_BUTTON = [
+    {
+      id: 1,
+      message: "모두 동의합니다.",
+    },
+    {
+      id: 2,
+      message: "[필수] FLAG 계정 약관",
+    },
+    {
+      id: 3,
+      message: "[필수] 개인정보 수집 및 이용 동의",
+    },
+  ];
