+++ src/constants/user.js
@@ -1,0 +1,21 @@
+// eslint-disable-next-line
+export const regExp = /[\{\}\[\]\/?.,;:|\)*~`!^\-_+<>@\#$%&\\\=\(\'\"]/g;
+export const korExp = /[ㄱ-ㅎㅏ-ㅣ가-힣]/g;
+export const numExp = /[0-9]/g;
+export const spaceExp = /\s/;
+export const engExp = /[a-zA-Z]/g;
+
+export const SPECIALIZED = [
+  {
+    label: "전공을 선택하세요",
+    value: "전공을 선택하세요",
+  },
+  {
+    label: "국어국문",
+    value: "국어국문",
+  },
+  {
+    label: "컴퓨터SW",
+    value: "컴퓨터SW",
+  },
+];
