+++ src/pages/board/EditPost.jsx
@@ -3,7 +3,6 @@ import { useSelector } from "react-redux";
 import styled from "styled-components";
 import boardAPI from '../../apis/boardAPI'; 
 import boardData from "../../constants/board";
-import buttonData from "../../constants/button";
 import { Footer, SideBar, Header } from "../../components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useNavigate } from "react-router-dom";
@@ -90,7 +89,7 @@ const EditPost = () => {
                             }}
                         />
                         <ContentButtonBox>
-                        {buttonData.BUTTON_ITEMS.map(({id, faIcon, text}) => (
+                        {boardData.FILE_IMAGE_BUTTON.map(({id, faIcon, text}) => (
                             <ContentButton key={id}>
                             <FaIcon icon={faIcon}/>
                             <p>{text}</p>
