+++ src/pages/index.js
@@ -1,12 +1,5 @@
-
-export {default as ChangePw} from "./ChangePw";
-export {default as EditUser} from "./EditUser";
-export {default as FindId} from "./FindId";
-export {default as FindPw} from "./FindPw";
-export {default as Home} from "./Home";
-export {default as LoginPage} from "./LoginPage";
-export {default as MyPage} from "./MyPage";
-export {default as Resume} from "./Resume";
-export {default as SearchPage} from "./SearchPage";
-export {default as SignUp} from "./SignUpPage";
-export {default as Activity} from "./Activity";
+export { default as ChangePw } from "./user/ChangePw";
+export { default as Home } from "./Home";
+export { default as Resume } from "./Resume";
+export { default as SearchPage } from "./SearchPage";
+export { default as Activity } from "./Activity";
