+++ src/pages/login/index.js
@@ -1,0 +1,4 @@
+export {default as FindId} from "./FindId";
+export {default as FindPw} from "./FindPw";
+export {default as LoginPage} from "./LoginPage";
+export {default as SignUpPage} from "./SignUpPage";
