+++ src/pages/user/index.js
@@ -1,0 +1,3 @@
+export {default as ChangePw} from "./ChangePw";
+export {default as EditUser} from "./EditUser";
+export {default as MyPage} from "./MyPage";
